Moved config and modified scripts
