+++ extension/readme/src/background.js
@@ -1,12 +1,12 @@
 'use strict';
-var {getArticleById } = require('./shared/network/apiHelper'); 
+var {getArticle } = require('../../../shared/network/lib/article'); 
 // With background scripts you can communicate with popup
 // and contentScript files.
 // For more information on background script,
 // See https://developer.chrome.com/extensions/background_pages
 
 chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
-  if (request.type === 'GREETINGS') {
+  if (request.type === 'ARTICLE') {
     const message = `Hi ${
       sender.tab ? 'Con' : 'Pop'
     }, my name is Bac. I am from Background. It's great to hear from you.`;
