+++ shared/network/apiClient.js
@@ -1,9 +1,9 @@
-const {config} = require('./config/prodConfig');
+const {config} = require('./config');
 const axios = require('axios');
 import 'regenerator-runtime/runtime'
 
 const axiosClient = axios.create({
-    baseUrl: config['db-host'],
+    baseUrl: `${config['host']}:${config['db-port']}`,
     headers: {
         'Accept': 'application/json',
         'Content-Type': 'application/json'
@@ -17,7 +17,7 @@ axiosClient.interceptors.response.use(
     function(error){
         let res = error.response;
         if(res.status == 401){
-            window.location.href = `${config["dashboard-host"]}/${config["loginPath"]}`
+            window.location.href = `${config['host']}:${config['dashboard-port']}/${config['loginPath']}`
         }
         console.error(`Looks like there was a problem. Status code: ${res.status}`);
         return Promise.reject;
