+++ shared/network/config.js
@@ -1,6 +1,7 @@
 export const config = {
-    "db-host": "http://20.71.92.74:8081",
-    "dashboard-host": "http://20.71.92.74:8080",
+    "host": `__API__` ,
+    "db-port": "8081",
+    "dashboard-port": "8080",
     "loginPath": "/login",
     "getArticlePath": "/api/getArticle",
     "getArticlesPath": "/api/getArticles",
