--- shared/network/config/devConfig.js
@@ -1,13 +1,0 @@
-export const config = {
-    "db-host": "http://127.0.0.1:8081",
-    "dashboard-host" : "https://127.0.0.1:8080",
-    "loginPath": "/login",
-    "getArticlePath": "/api/getArticle",
-    "getArticlesPath": "/api/getArticles",
-    "getUserPath": "/api/getUser",
-    "getUsersPath": "/api/getUsers",
-    "newUserPath": "/api/newUser",
-    "newArticle": "/api/newArticle",
-    "updateUser": "/api/updateUser",
-    "updateArticle" : "/api/updateArticle"
-}
